# refactor(meta): split meta into smaller crates (step 1)