--- a/src/compute/src/lib.rs
+++ b/src/compute/src/lib.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #![feature(trait_alias)]
-#![feature(binary_heap_drain_sorted)]
 #![feature(generators)]
 #![feature(type_alias_impl_trait)]
 #![feature(let_chains)]