--- a/src/meta/service/src/health_service.rs
+++ b/src/meta/service/src/health_service.rs
@@ -19,6 +19,12 @@ use tonic::{Request, Response, Status};
 
 pub struct HealthServiceImpl {}
 
+impl Default for HealthServiceImpl {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl HealthServiceImpl {
     pub fn new() -> Self {
         Self {}