I hereby agree to the terms of the [RisingWave Labs, Inc. Contributor License Agreement](https://gist.github.com/TennyZhuang/f00be7f16996ea48effb049aa7be4d66#file-rw_cla).

## What's changed and what's your intention?

Just trying my idea quickly. Feel free to leave any comments/concerns. Will proceed if the idea sounds good.

I came up with this idea because I thought `async_trait` is a major reason of the slow compilation (https://github.com/risingwavelabs/risingwave/issues/9553#issuecomment-1766602371). There are many `async_trait`s in tonic services, and many heavy futures to type check. We can hardly get rid of them, but maybe we can split the services into crates so that we can compile them in parallel. (a little similar to #12485, but different)

As a first step, split out:
- `meta/service`: tonic grpc service implementations. We may further split this into parallel sub-crates.
- `meta/node`: the server binary

### Result
A rough manual test:
compiling only meta takes 50s -> 42s
compiling pb -> cmd_all takes 92s -> 76s

Before
<img width="1492" alt="image" src="https://github.com/risingwavelabs/risingwave/assets/37948597/ccaec025-d33c-4542-8d07-b7867f2a7011">

After
<img width="1492" alt="image" src="https://github.com/risingwavelabs/risingwave/assets/37948597/9ee9a07f-a189-4e48-9cbe-b3e836d911c8">



### Note
In order to quickly pass compilation, some brute force is used:
- All `pub(crate)` are simply replaced with `pub`.
- `use risingwave_meta::*` is used without careful fine-graind imports.

<!--

**Please do not leave this empty!**

Please explain **IN DETAIL** what the changes are in this PR and why they are needed:

- Summarize your change (**mandatory**)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Describe clearly one logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)
- Refer to a related PR or issue link (optional)

-->

## Checklist

- [ ] I have written necessary rustdoc comments
- [ ] I have added necessary unit tests and integration tests
- [ ] I have added fuzzing tests or opened an issue to track them. (Optional, recommended for new SQL features #7934).
- [ ] My PR contains breaking changes. (If it deprecates some features, please create a tracking issue to remove them in the future).
- [ ] All checks passed in `./risedev check` (or alias, `./risedev c`)
- [ ] My PR changes performance-critical code. (Please run macro/micro-benchmarks and show the results.)
<!-- To manually trigger a benchmark, please check out [Notion](https://www.notion.so/risingwave-labs/Manually-trigger-nexmark-performance-dashboard-test-b784f1eae1cf48889b2645d020b6b7d3). -->
- [ ] My PR contains critical fixes that are necessary to be merged into the latest release. (Please check out the [details](https://github.com/risingwavelabs/risingwave/blob/main/CONTRIBUTING.md))

## Documentation

- [ ] My PR needs documentation updates. (Please use the **Release note** section below to summarize the impact on users)

## Release note

If this PR includes changes that directly affect users or other significant modifications relevant to the community, kindly draft a release note to provide a concise summary of these changes. Please prioritize highlighting the impact these changes will have on users.


<!--
Please create a release note for your changes.

Discuss technical details in the "What's changed" section, and
focus on the impact on users in the release note.

You should also mention the environment or conditions where the impact may occur.
-->

</details>