@xxchan 2023-10-18
> How were you able to do this so quickly?

🤔️ It took me about 2h to do this. Steps were:

1. Simply mv `meta/src/rpc/service`, to `meta_service` and `meta/src/lib.rs`, `meta/src/rpc/server.rs` to `meta_node`. 
    - `Cargo.toml` is copied from meta (without pruning the dependencies), and add meta as dependency.
    - Add `use risingwave_meta::*` to new crates so that we don't need to change `use crate::..`.
2. make `cargo build -p risingwave_meta` compile. IIRC this is very easy, thanks to the fact that the dependency relationship of these are very clear.
3. make `cargo build -p risingwave_meta_service` compile. Mainly just need to change `pub(crate)` to `pub`... Then fix a few minor issues. IIRC this part took the most time.
4. make `cargo build -p risingwave_meta_node` compile. Easier. Mainly just replace imports.
5. make `cargo build -p risingwave_cmd_all` compile.
6. `risedev check-fix`, and fix a few minor issues.

As you can see the diff of this PR is also very reviewable, so it's a relatively easy refactor. I think it should mainly because the dependency `node <- service <- others` is clear.